Failed try: load the DefaultLegacyService into its own class loader
This would have been so cool: by having an own class loader for each
DefaultLegacyService instance, we would have let ImageJ 1.x continue to
pretend that it is a singleton, but it would not really have been (since
its class definition would have been local to the legacy service).

Alas, it is not to be. If we were to do that, nobody could instantiate an
ImagePlus *except* in classes loaded by those class loaders, which would
have been unintuitive and fragile. So a simple

	ImagePlus image = IJ.openImage(path);
	ImageDisplay display =
		legacyService.getImageMap().lookupDisplay(image);

would have thrown a non-intuitive ClassCastException, stating that
"ij.ImagePlus could not be cast to ij.ImagePlus" because the respective
classes' class definitions would be incompatible by virtue of being
defined in different class loaders.

So we have to undo this change, recording for history that we tried.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>